+++ src/Raven.Server/Documents/Patch/PatchDocument.cs
@@ -144,8 +144,8 @@ namespace Raven.Server.Documents.Patch
                 var isModified = document.Data.Size != modifiedDocument.Size;
                 if (isModified == false) // optimization, if size different, no need to compute hash to check
                 {
-                    var originHash = Hashing.XXHash64.Calculate(document.Data.BasePointer, document.Data.Size);
-                    var modifiedHash = Hashing.XXHash64.Calculate(modifiedDocument.BasePointer, modifiedDocument.Size);
+                    var originHash = Hashing.XXHash64.Calculate(document.Data.BasePointer, (ulong)document.Data.Size);
+                    var modifiedHash = Hashing.XXHash64.Calculate(modifiedDocument.BasePointer, (ulong)modifiedDocument.Size);
                     isModified = originHash != modifiedHash;
                 }
